Fix issues in bhashini utils implementation
